7_requirements.txt
absl-py==1.4.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
awscli==1.27.68
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
boto3==1.26.68
botocore==1.29.68
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
brotlipy==0.7.0
cachetools==5.2.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer==3.0.1
colorama==0.4.4
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
cryptography @ file:///C:/b/abs_b7d7drzbky/croot/cryptography_1673298763653/work
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distro==1.8.0
docopt==0.6.2
docutils==0.16
entrypoints @ file:///C:/ci/entrypoints_1649926676279/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
flatbuffers==23.1.4
gast==0.4.0
git-filter-repo==2.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h5py==3.7.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
ipykernel @ file:///C:/b/abs_b4f07tbsyd/croot/ipykernel_1672767104060/work
ipython @ file:///C:/b/abs_536xsn7rj0/croot/ipython_1670919360443/work
jedi @ file:///C:/ci/jedi_1644315428305/work
jmespath==1.0.1
jupyter_client @ file:///C:/b/abs_b9pns5mx5p/croot/jupyter_client_1671703062216/work
jupyter_core @ file:///C:/b/abs_84df679bho/croot/jupyter_core_1672332237650/work
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934094726/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging @ file:///C:/b/abs_cfsup8ur87/croot/packaging_1671697442297/work
pandas @ file:///C:/b/abs_c6fuezktfm/croot/pandas_1670425103552/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pipreqs==0.4.11
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.19.6
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
psycopg2==2.9.5
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/b/abs_22fofvpn1x/croot/pytz_1671698059864/work
pywin32==305.1
PyYAML==5.4.1
pyzmq @ file:///C:/ci/pyzmq_1657616000714/work
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.0
selenium @ file:///C:/ci_310/selenium_1642091463950/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==1.4.46
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabula-py==2.6.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tornado @ file:///C:/ci/tornado_1662476985533/work
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
typing_extensions==4.4.0
urllib3 @ file:///C:/b/abs_21qa0j12xt/croot/urllib3_1670527000672/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.2
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
wrapt==1.14.1
yarg==0.1.9
